Numeric Delta Method approximation for the variance-covariance matrix

Computes delta method variance-covariance matrix of results of any generic function fct that computes a vector of estimates as a function of a set of estimated parameters par.

Usage

DeltaMethod(par, fct, vcov, delta, ...)

Arguments

par

vector of parameter values at which estimates should be constructed

fct

function that constructs estimates from parameters par

vcov

variance-covariance matrix of the parameters

delta

proportional change in parameters used to numerically estimate first derivative with central-difference formula (ignored)

...

any additional arguments needed by fct

Value

a list with values

variance

estimated variance-covariance matrix of estimates derived by fct

partial

matrix (or vector) of partial derivatives of fct with respect to the parameters par

Details

The delta method (aka propagation of errors is based on Taylor series approximation - see Seber's book on Estimation of Animal Abundance). It uses the first derivative of fct with respect to par. It also uses the variance-covariance matrix of the estimated parameters which is derived in estimating the parameters and is an input argument.

The first argument of fct should be par which is a vector of parameter estimates. It should return a single value (or vector) of estimate(s). The remaining arguments of fct if any can be passed to fct by including them at the end of the call to DeltaMethod as name=value pairs.

Note

This is a generic function that can be used in any setting beyond the mrds package. However this is an internal function for mrds and the user does not need to call it explicitly.

Author

Jeff Laake and David L Miller