Series of the gradient of the Hermite polynomial adjustment series w.r.t. the scaled distance.

For internal use only – not to be called by 'mrds' or 'Distance' users directly.

Usage

adj.series.grad.herm(
  distance,
  scaling = 1,
  adj.order,
  adj.parm = NULL,
  adj.exp = FALSE
)

Arguments

distance

perpendicular distance vector/scalar

scaling

scale parameter

adj.order

the adjustment order

adj.parm

vector of parameters (a_j)

adj.exp

boolean, defaults to FALSE

Value

scalar or vector containing the gradient of the Hermite adjustment series for every value in distance argument

Author

Felix Petersma