Golf tee data used in chapter 6 of Advanced Distance Sampling examples

Double platform data collected in a line transect survey of golf tees by 2 observers at St. Andrews. Field sex was actually colour of the golf tee: 0 - green; 1 - yellow. Exposure was either low (0) or high(1) depending on height of tee above the ground. size was the number of tees in an observed cluster.

Format

A list of 4 dataframes, with the list elements named: book.tee.dataframe, book.tee.region, book.tee.samples and book.tee.obs.

book.tee.dataframe is the distance sampling data dataframe. Used in the call to fit the detection function in ddf. Contains the following columns:

object

numeric object id

observer

factor representing observer 1 or 2

detected

numeric 1 if the animal was detected 0 otherwise

distance

numeric value for the distance the animal was detected

size

numeric value for the group size

sex

numeric value for sex of animal

exposure

numeric value for exposure level 0 or 1

book.tee.region: is the region table dataframe. Used to supply the strata areas to the dht function. Contains the following columns:

Region.Label

factor giving the strata labels

Area

numeric value giving the strata areas

book.tee.samples is the samples table dataframe to match the transect ids to the region ids and supply the effort. Used in the dht function. Contains the following columns:

Sample.Label

numeric giving the sample / transect labels

Region.Label

factor giving the strata labels

Effort

numeric value giving the sample / transect lengths

book.tee.obs is the observations table dataframe to match the object ids in the distance data to the transect labels. Used in the dht function. Contains the following columns:

object

numeric value object id

Region.Label

factor giving the strata labels

Sample.Label

numeric giving the sample / transect labels