Extract coefficients

Extract coefficients and provide a summary of parameters and estimates from the output of ddf model objects.

Usage

# S3 method for class 'ds'
coef(object,...)
       # S3 method for class 'io'
coef(object,...)
       # S3 method for class 'io.fi'
coef(object,...)
       # S3 method for class 'trial'
coef(object,...)
       # S3 method for class 'trial.fi'
coef(object,...)
       # S3 method for class 'rem'
coef(object,...)
       # S3 method for class 'rem.fi'
coef(object,...)

Arguments

object

ddf model object of class ds, io, io.fi, trial, trial.fi, rem, or rem.fi.

...

unspecified arguments that are unused at present

Value

For coef.ds List of data frames for coefficients (scale and exponent (if hazard))

scale

dataframe of scale coefficient estimates and standard errors

exponent

dataframe with exponent estimate and standard error if hazard detection function

For all others Data frame containing each coefficient and standard error

Note

These functions are called by the generic function coef for any ddf model object. It can be called directly by the user, but it is typically safest to use coef which calls the appropriate function based on the type of model.

Author

Jeff Laake