Detection functions

Various functions used to specify key and adjustment functions for detection functions.

Usage

detfct(distance, ddfobj, select=NULL, index=NULL, width=NULL,
              standardize = TRUE, stdint=FALSE, left=0)

adjfct.cos(distance, scaling = 1, adj.order, adj.parm = NULL, adj.exp=FALSE)

adjfct.poly(distance, scaling = 1, adj.order, adj.parm = NULL, adj.exp=FALSE)

adjfct.herm(distance, scaling = 1, adj.order, adj.parm = NULL, adj.exp=FALSE)

scalevalue(key.scale, z)

keyfct.hn(distance, key.scale)

keyfct.hz(distance, key.scale, key.shape)

keyfct.gamma(distance, key.scale, key.shape)

fx(distance,ddfobj,select=NULL,index=NULL,width=NULL,
   standardize=TRUE,stdint=FALSE, left=0)

fr(distance,ddfobj,select=NULL,index=NULL,width=NULL,
   standardize=TRUE,stdint=FALSE)

distpdf(distance,ddfobj,select=NULL,index=NULL,width=NULL,standardize=TRUE,
           stdint=FALSE,point=FALSE, left=0)

Arguments

distance

vector of distances

ddfobj

distance sampling object (see create.ddfobj)

select

logical vector for selection of data values

index

specific data row index

width

(right) truncation width

standardize

logical used to decide whether to divide through by the function evaluated at 0

stdint

logical used to decide whether integral is standardized

point

if TRUE, point counts; otherwise line transects

left

(left) truncation distance

z

design matrix for scale function

key.scale

vector of scale values

key.shape

vector of shape values

adj.order

vector of adjustment orders

adj.parm

vector of adjustment parameters

scaling

the scaling for the adjustment terms

adj.exp

if TRUE uses exp(adj) for adjustment to keep f(x)>0

Value

For detfct, the value is a vector of detection probabilities For keyfct.*, vector of key function evaluations For adjfct.*, vector of adjustment series evaluations For scalevalue, vector of the scale parameters.

Details

Multi-covariate detection functions (MCDS) are represented by a function \(g(x,w,\theta)\) where x is distance, z is a set of covariates and \(\theta\) is the parameter vector. The functions are defined such that \(g(0,w,\theta)=1\) and the covariates modify the scale \((x/\sigma)\) where a log link is used to relate \(\sigma\) to the covariates, \(\sigma=exp(\theta*w)\). A CDS function is obtained with a constant \(\sigma\) which is equivalent to an intercept design matrix, z.

detfct will call either a gamma, half-normal, hazard-rate or uniform function only returning the probability of detection at that distance. In addition to the simple model above, we may specify adjustment terms to fit the data better. These adjustments are either Cosine, Hermite and simple polynomials. These are specified as arguments to detfct, as detailed below.

detfct function which calls the others and assembles the final result using either key(x)[1+series(x)] or (key(x)[1+series(x)])/(key(0)[1+series(0)]) (depending on the value of standardize).

keyfct.* functions calculate key function values and adjfct.* calculate adjustment term values.

scalevalue for either detection function it computes the scale with the log link using the parameters and the covariate design matrix

fx, fr non-normalized probability density for line transects and point counts respectively

References

Marques, F. F. C., & Buckland, S. T. (2003). Incorporating covariates into standard line transect analyses. Biometrics, 59(4), 924-935.

Buckland, S. T., Anderson, D. R., Burnham, K. P., Laake, J. L., Borchers, D. L., & Thomas, L. (2004). Advanced Distance Sampling. Oxford University Press, Oxford, UK.

Becker, E. F. and P. X. Quang. 2009. A gamma-shaped detection function for line transect surveys with mark-recapture and covariate data. Journal of Agricultural Biological and Environmental Statistics 14:207-223.

See also

Author

Jeff Laake, David L Miller