Distance Sampling Functions

Computes values of conditional and unconditional detection functions and probability density functions for for line/point data for single observer or dual observer in any of the 3 configurations (io,trial,rem).

Usage

ds.function(
  model,
  newdata = NULL,
  obs = "All",
  conditional = FALSE,
  pdf = TRUE,
  finebr
)

Arguments

model

model object

newdata

dataframe at which to compute values; if NULL uses fitting data

obs

1 or 2 for observer 1 or 2, 3 for duplicates, "." for combined and "All" to return all of the values

conditional

if FALSE, computes p(x) based on distance detection function and if TRUE based on mr detection function

pdf

if FALSE, returns p(x) and if TRUE, returns p(x)*pi(x)/integral p(x)*pi(x)

finebr

fine break values over which line is averaged

Value

List containing

xgrid

grid of distance values

values

average detection fct values at the xgrid values

Details

Placeholder – Not functional —-

Author

Jeff Laake