Integral of pdf of distances

Computes the integral of distpdf with scale=1 (stdint=TRUE) or specified scale (stdint=FALSE).

Usage

gstdint(
  x,
  ddfobj,
  index = NULL,
  select = NULL,
  width,
  standardize = TRUE,
  point = FALSE,
  stdint = TRUE,
  doeachint = FALSE,
  left = left
)

Arguments

x

lower, upper value for integration

ddfobj

distance detection function specification

index

specific data row index

select

logical vector for selection of data values

width

truncation width

standardize

if TRUE, divide through by the function evaluated at 0

point

logical to determine if point (TRUE) or line transect(FALSE)

stdint

if TRUE, scale=1 otherwise specified scale used

doeachint

if TRUE perform integration using integrate

left

left truncation width

Value

vector of integral values of detection function

Note

This is an internal function that is not intended to be invoked directly.

Author

Jeff Laake and David L Miller