Plot histogram line

Takes bar heights (height) and cutpoints (breaks), and constructs a line-only histogram from them using the function plot() (if lineonly==FALSE) or lines() (if lineonly==TRUE).

Usage

histline(
  height,
  breaks,
  lineonly = FALSE,
  outline = FALSE,
  ylim = range(height),
  xlab = "x",
  ylab = "y",
  det.plot = FALSE,
  add = FALSE,
  ...
)

Arguments

height

heights of histogram bars

breaks

cutpoints for x

lineonly

if TRUE, drawn with plot; otherwise with lines to allow addition of current plot

outline

if TRUE, only outline of histogram is plotted

ylim

limits for y axis

xlab

label for x axis

ylab

label for y axis

det.plot

if TRUE, plot is of detection so yaxis limited to unit interval

add

should this plot add to a previous window

...

Additional unspecified arguments for plot

Value

None

Author

Jeff Laake and David L Miller