MCDS function definition

Creates model formula list for multiple covariate distance sampling using values supplied in call to ddf

Usage

mcds(
  formula = NULL,
  key = NULL,
  adj.series = NULL,
  adj.order = c(NULL),
  adj.scale = "width",
  adj.exp = FALSE,
  shape.formula = ~1
)

Arguments

formula

formula for scale function

key

string identifying key function (currently either "hn" (half-normal),"hr" (hazard-rate), "unif" (uniform) or "gamma" (gamma distribution)

adj.series

string identifying adjustment functions cos (Cosine), herm (Hermite polynomials), poly (simple polynomials) or NULL

adj.order

vector of order of adjustment terms to include

adj.scale

whether to scale the adjustment terms by "width" or "scale"

adj.exp

if TRUE uses exp(adj) for adjustment to keep f(x)>0

shape.formula

formula for shape function

Value

A formula list used to define the detection function model

fct

string "mcds"

key

key function string

adj.series

adjustment function string

adj.order

adjustment function orders

adj.scale

adjustment function scale type

formula

formula for scale function

shape.formula

formula for shape function

Author

Jeff Laake; Dave Miller