Plot unconditional detection function from distance sampling model

Plots unconditional detection function for observer=obs observations overlays histogram, average detection function and values for individual observations data. Internal function called by plot methods.

Usage

plot_uncond(
  model,
  obs,
  xmat,
  gxvalues,
  nc,
  finebr,
  breaks,
  showpoints,
  showlines,
  maintitle,
  ylim,
  return.lines = FALSE,
  angle = -45,
  density = 20,
  col = "black",
  jitter = NULL,
  xlab = "Distance",
  ylab = "Detection probability",
  subtitle = TRUE,
  ...
)

Arguments

model

fitted model from ddf

obs

value of observer for plot

xmat

processed data

gxvalues

detection function values for each observation

nc

number of equal-width bins for histogram

finebr

fine break values over which line is averaged

breaks

user define breakpoints

showpoints

logical variable; if TRUE plots predicted value for each observation

showlines

logical variable; if TRUE plots average predicted value line

maintitle

main title line for each plot

ylim

range of y axis; defaults to (0,1)

return.lines

if TRUE, returns values for line

angle

shading angle for hatching

density

shading density for hatching

col

plotting colour

jitter

scaling option for plotting points. Jitter is applied to points by multiplying the fitted value by a random draw from a normal distribution with mean 1 and sd jitter.

xlab

label for x-axis

ylab

label for y-axis

subtitle

if TRUE, shows plot type as sub-title

...

other graphical parameters, passed to the plotting functions (plot, hist, lines, points, etc)

Value

if return.lines==TRUE returns dataframe average.line

otherwise just plots

Author

Jeff Laake, Jon Bishop, David Borchers