Derivatives for variance of average p and average p(0) variance

Used in call to DeltaMethod from prob.se to get first derivatives

Usage

prob.deriv(par, model, parfct, observer = NULL, fittedmodel = NULL)

Arguments

par

detection function parameter values

model

ddf model object

parfct

function of detection probabilities; currently only average (over covariates) detection probability p integrated over distance or average (over covariates) detection probability at distance 0; p(0)

observer

1,2,3 for primary, secondary, or duplicates for average p(0); passed to fct

fittedmodel

full fitted ddf model when trial.fi or io.fi is called from trial or io respectively

Value

Vector of values from fct at specified parameter values

Details

Need to add equations here as I do not think they exist in any of the texts. These should probably be checked with simulation.

See also

Author

Jeff Laake