Average p and average p(0) variance

Computes components of variance for average p=n/N and average p(0) with weights based on empirical covariate distribution, if it contains covariates.

Usage

prob.se(model, fct, vcov, observer = NULL, fittedmodel = NULL)

Arguments

model

ddf model object

fct

function of detection probabilities; currently only average (over covariates) detection probability p integrated over distance or average (over covariates) detection probability at distance 0; p(0)

vcov

variance-covariance matrix of parameter estimates

observer

1,2,3 for primary, secondary, or duplicates for average p(0); passed to fct

fittedmodel

full fitted ddf model when trial.fi or io.fi is called from trial or io respectively

Value

var

variance

partial

partial derivatives of parameters with respect to fct

covar

covariance of n and average p or p(0)

Details

Need to add equations here as I do not think they exist in any of the texts. These should probably be checked with simulation.

See also

Author

Jeff Laake