Calculate the parameter rescaling for parameters associated with covariates

This will calculate the rescaling needed when covariates to be included in the scale of the detection function are "too big". Based on code from optimx.

Usage

rescale_pars(initialvalues, ddfobj)

Arguments

initialvalues

starting values for the optimisation

ddfobj

detection function object

Details

Derivative-free methods like nlminb are sensitive to the parameters being poorly scaled. This can also cause problems for quasi-Newton methods too (at least, bad scaling won't _help_ the optimisation). So here we rescale the parameters if necessary (unless we already got scaling from control)

Author

David L Miller