Summary of distance detection function model object

Provides a brief summary of data and fitted detection probability model parameters, model selection criterion, and optionally abundance in the covered (sampled) region and its standard error.

Usage

# S3 method for io
summary(object, se = TRUE, ...)

Arguments

object

a ddf model object

se

if TRUE, computes standard errors

...

unspecified and unused arguments for S3 consistency

Value

list of extracted and summarized objects

Details

The argument N is used to suppress computation of abundance and average detection probability in calls to summarize the ds and either io.fi or trial.fi for summaries of io and trial objects respectively which are composed of a ds model object and a mark-recapture model object. The corresponding print function is called to print the summary results.

Note

This function is called by the generic function summary for any ddf model object. Each function can be called directly by the user, but it is typically safest to use the generic function summary which calls the appropriate function based on the type of ddf model.

Author

Jeff Laake