Variance and confidence intervals for density and abundance estimates

Computes standard error, cv, and log-normal confidence intervals for abundance and density within each region (if any) and for the total of all the regions. It also produces the correlation matrix for regional and total estimates.

Usage

dht.se(
  model,
  region.table,
  samples,
  obs,
  options,
  numRegions,
  estimate.table,
  Nhat.by.sample
)

Arguments

model

ddf model object

region.table

table of region values

samples

table of samples(replicates)

obs

table of observations

options

list of options that can be set (see dht)

numRegions

number of regions

estimate.table

table of estimate values

Nhat.by.sample

estimated abundances by sample

Value

List with 2 elements:

estimate.table

completed table with se, cv and confidence limits

vc

correlation matrix of estimates

Details

The variance has two components:

variation due to uncertainty from estimation of the detection function parameters;

variation in abundance due to random sample selection.

The first component (model parameter uncertainty) is computed using a delta method estimate of variance (Huggins 1989 ; Huggins 1991 ; Borchers et al. 1998 ) in which the first derivatives of the abundance estimator with respect to the parameters in the detection function are computed numerically (see DeltaMethod).

The second component (encounter rate variance) can be computed in one of several ways depending on the form taken for the encounter rate and the estimator used. To begin with there three possible values for varflag to calculate encounter rate:

0 uses a negative binomial variance for the number of observations (equation 13 of Borchers et al. 1998 ). This estimator is only useful if the sampled region is the survey region and the objects are not clustered; this situation will not occur very often;

1 uses the encounter rate \(n/L\) (objects observed per unit transect) from Buckland et al. (2001) pg 78-79 (equation 3.78) for line transects (see also Fewster et al. 2009 estimator R2). This variance estimator is not appropriate if size or a derivative of size is used in the detection function;

2 is the default and uses the encounter rate estimator \(\hat{N}/L\) (estimated abundance per unit transect) suggested by Innes et al. (2002) and Marques and Buckland (2004) .

In general if any covariates are used in the models, the default varflag=2 is preferable as the estimated abundance will take into account variability due to covariate effects. If the population is clustered the mean group size and standard error is also reported.

For options 1 and 2, it is then possible to choose one of the estimator forms given in Fewster et al. (2009) . For line transects: "R2", "R3", "R4", "S1", "S2", "O1", "O2" or "O3" can be used by specifying ervar in the list of options provided by the options argument (default "R2"). For points, either the "P2" or "P3" estimator can be selected (>=mrds 2.3.0 default "P2", <= mrds 2.2.9 default "P3"). See varn and Fewster et al. (2009) for further details on these estimators.

Exceptions to the above occur if there is only one sample in a stratum. In this situation, varflag=0 continues to use a negative binomial variance while the other options assume a Poisson variance (\(Var(x)=x\)), where when varflag=1 x is number of detections in the covered region and when varflag=2 x is the abundance in the covered region. It also assumes a known variance so \(z=1.96\) is used for critical value. In all other cases the degrees of freedom for the \(t\)-distribution assumed for the log(abundance) or log(density) is based on the Satterthwaite approximation (Buckland et al. 2001 pg 90) for the degrees of freedom (df). The df are weighted by the squared cv in combining the two sources of variation because of the assumed log-normal distribution because the components are multiplicative. For combining df for the sampling variance across regions they are weighted by the variance because it is a sum across regions.

The coefficient of variation (CV) associated with the abundance estimates is calculated based on the following formula for the varflag options 1 and 2:

varflag=1

$$CV(\hat{N}) = \sqrt{\left(\frac{\sqrt{n}}{n}\right)^2+CV(\hat{p})^2}$$

varflag=2

$$CV(\hat{N}) = \sqrt{\left(\frac{\sqrt{\hat{N}}}{\hat{N}}\right)^2+CV(\hat{p})^2}$$ where n is the number of observations, \(\hat{N}\) is the estimated abundance and \(\hat{p}\) is the average probability of detection for an animal in the covered area.

A non-zero correlation between regional estimates can occur from using a common detection function across regions. This is reflected in the correlation matrix of the regional and total estimates which is given in the value list. It is only needed if subtotals of regional estimates are needed.

Note

This function is called by dht and it is not expected that the user will call this function directly but it is documented here for completeness and for anyone expanding the code or using this function in their own code.

References

Borchers DL, Buckland ST, Goedhart PW, Clarke ED, Hedley SL (1998). “Horvitz-Thompson Estimators for Double-Platform Line Transect Surveys.” Biometrics, 54(4), 1221-1237. doi:10.2307/253365 .

Buckland ST, Anderson DR, Burnham KP, Laake JL, Borchers DL, Thomas L (2001). Introduction to distance sampling: estimating abundance of biological populations. Oxford university press.

Fewster RM, Buckland ST, Burnham KP, Borchers DL, Jupp PE, Laake JL, Thomas L (2009). “Estimating the encounter rate variance in distance sampling.” Biometrics, 65(1), 225-236.

Huggins RM (1989). “On the statistical analysis of capture experiments.” Biometrika, 76(1), 133-140. doi:10.1093/biomet/76.1.133 .

Huggins RM (1991). “Some practical aspects of a conditional likelihood approach to capture experiments.” Biometrics, 47(1), 725-732. doi:10.1093/biomet/76.1.133 .

Innes S, Heide-Jørgensen MP, Laake JL, Laidre KL, Cleator HJ, Richard P, Stewart RE (2002). “Surveys of belugas and narwhals in the Canadian High Arctic in 1996.” NAMMCO Scientific Publications, 4, 169-190.

Marques FFC, Buckland ST (2004). “Advanced distance sampling.” In chapter Covariate models for the detection function, 31-47. Oxford University Press.

See also

Author

Jeff Laake