Compute empirical variance of encounter rate

Computes one of a series of possible variance estimates for the observed encounter rate for a set of sample measurements (e.g., line lengths) and number of observations per sample.

Usage

varn(lvec,nvec,type)

         covn(lvec, groups1, groups2, type)

Arguments

lvec

vector of sample measurements (e.g., line lengths)

nvec

vector of number observed

type

choice of variance estimator to use for encounter rate

groups1

vector of number of groups observed

groups2

vector of number of individuals observed

Value

Variance of encounter rate as defined by arguments

Details

The choice of type follows the notation of Fewster et al. (2009) in that there are 8 choices of encounter rate variance that can be computed for lines and one for points:

R2

random line placement with unequal line lengths (design-assisted estimator)

R3

random line placement, model-assisted estimator, based on true contagion process

R4

random line placement, model-assisted estimator, based on apparent contagion process

S1

systematic line placement, post-stratification with no strata overlap

S2

systematic line placement, post-stratification with no strata overlap, variances weighted by line length per stratum

O1

systematic line placement, post-stratification with overlapping strata (akin to S1)

O2

systematic line placement, post-stratification with overlapping strata (weighted by line length per stratum, akin to S2)

O3

systematic line placement, post-stratification with overlapping strata, model-assisted estimator with trend in encounter rate with line length

P2

random point placement, potentially unequal number of visits per point, design-based estimator

P3

random point placement, potentially unequal number of visits per point, model-based estimator

Default value is "R2", shown in Fewster et al. (2009) to have good performance for completely random designs for lines. For systematic parallel line transect designs, Fewster et al. (2009) recommend "O2". For point transects the default is "P2" (but "P3" is also available).

For the systematic estimators, pairs are assigned in the order they are given in the lengths and groups vectors.

Note

This function is also used with different calling arguments to compute Innes et al. (2002) variance of the estimated abundances/length rather than observation encounter rate. The function covn is probably only valid for R3 and R2. Currently, the R2 form is used for all types other than R3.

References

Fewster RM, Buckland ST, Burnham KP, Borchers DL, Jupp PE, Laake JL, Thomas L (2009). “Estimating the encounter rate variance in distance sampling.” Biometrics, 65(1), 225-236.

Innes S, Heide-Jørgensen MP, Laake JL, Laidre KL, Cleator HJ, Richard P, Stewart RE (2002). “Surveys of belugas and narwhals in the Canadian High Arctic in 1996.” NAMMCO Scientific Publications, 4, 169-190.

Author

Jeff Laake, David L Miller